+++ tools/fuzzer/topology.c
@@ -20,7 +20,9 @@
 
 #include "../fuzzer/fuzzer.h"
 
-const struct sof_dai_types sof_dais[] = {
+struct ipc;
+
+static const struct sof_dai_types sof_dais[] = {
 	{"SSP", SOF_DAI_INTEL_SSP},
 	{"HDA", SOF_DAI_INTEL_HDA},
 	{"DMIC", SOF_DAI_INTEL_DMIC},
@@ -41,6 +43,31 @@ enum sof_ipc_dai_type find_dai(const char *name)
 
 void register_comp(int comp_type, struct sof_ipc_comp_ext *comp_ext) {}
 
+int ipc_pipeline_complete(struct ipc *ipc, uint32_t comp_id)
+{
+	return 0;
+}
+
+int ipc_comp_connect(struct ipc *ipc, struct sof_ipc_pipe_comp_connect *_connect)
+{
+	return 0;
+}
+
+int ipc_buffer_new(struct ipc *ipc, const struct sof_ipc_buffer *desc)
+{
+	return 0;
+}
+
+int ipc_pipeline_new(struct ipc *ipc, struct sof_ipc_pipe_new *_pipe_desc)
+{
+	return 0;
+}
+
+int ipc_comp_new(struct ipc *ipc, struct sof_ipc_comp *_comp)
+{
+	return 0;
+}
+
 int find_widget(struct comp_info *temp_comp_list, int count, char *name)
 {
 	int i;
@@ -53,7 +80,7 @@ int find_widget(struct comp_info *temp_comp_list, int count, char *name)
 	return -EINVAL;
 }
 
-int complete_pipeline(struct fuzz *fuzzer, uint32_t comp_id)
+static int fuzzer_complete_pipeline(struct fuzz *fuzzer, uint32_t comp_id)
 {
 	struct sof_ipc_pipe_ready ready;
 	struct sof_ipc_reply r;
@@ -79,7 +106,7 @@ int complete_pipeline(struct fuzz *fuzzer, uint32_t comp_id)
 }
 
 /* load pipeline graph DAPM widget*/
-static int load_graph(void *dev, struct comp_info *temp_comp_list,
+static int fuzzer_load_graph(void *dev, struct comp_info *temp_comp_list,
 		      int count, int num_comps, int pipeline_id)
 {
 	struct sof_ipc_pipe_comp_connect connection;
@@ -90,7 +117,7 @@ static int load_graph(void *dev, struct comp_info *temp_comp_list,
 	int i;
 
 	for (i = 0; i < count; i++) {
-		ret = tplg_load_graph(num_comps, pipeline_id, temp_comp_list,
+		ret = tplg_create_graph(num_comps, pipeline_id, temp_comp_list,
 				      pipeline_string, &connection,
 				      fuzzer->tplg_file, i, count);
 		if (ret < 0)
@@ -111,14 +138,14 @@ static int load_graph(void *dev, struct comp_info *temp_comp_list,
 }
 
 /* load buffer DAPM widget */
-int load_buffer(struct tplg_context *ctx)
+static int fuzzer_load_buffer(struct tplg_context *ctx)
 {
 	struct sof_ipc_buffer buffer;
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_reply r;
 	int ret;
 
-	ret = tplg_load_buffer(ctx, &buffer);
+	ret = tplg_create_buffer(ctx, &buffer);
 	if (ret < 0)
 		return ret;
 
@@ -137,14 +164,14 @@ int load_buffer(struct tplg_context *ctx)
 }
 
 /* load pcm component */
-static int load_pcm(struct tplg_context *ctx, int dir)
+static int fuzzer_load_pcm(struct tplg_context *ctx, int dir)
 {
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_host host;
 	struct sof_ipc_comp_reply r;
 	int ret;
 
-	ret = tplg_load_pcm(ctx, dir, &host);
+	ret = tplg_create_pcm(ctx, dir, &host);
 	if (ret < 0)
 		return ret;
 
@@ -163,18 +190,18 @@ static int load_pcm(struct tplg_context *ctx, int dir)
 
 int load_aif_in_out(struct tplg_context *ctx, int dir)
 {
-	return load_pcm(ctx, dir);
+	return fuzzer_load_pcm(ctx, dir);
 }
 
 /* load dai component */
-static int load_dai(struct tplg_context *ctx)
+static int fuzzer_load_dai(struct tplg_context *ctx)
 {
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_dai comp_dai;
 	struct sof_ipc_comp_reply r;
 	int ret;
 
-	ret = tplg_load_dai(ctx, &comp_dai);
+	ret = tplg_create_dai(ctx, &comp_dai);
 	if (ret < 0)
 		return ret;
 
@@ -194,18 +221,18 @@ static int load_dai(struct tplg_context *ctx)
 
 int load_dai_in_out(struct tplg_context *ctx, int dir)
 {
-	return load_dai(ctx);
+	return fuzzer_load_dai(ctx);
 }
 
 /* load pda dapm widget */
-int load_pga(struct tplg_context *ctx)
+static int fuzzer_load_pga(struct tplg_context *ctx)
 {
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_volume volume;
 	struct sof_ipc_comp_reply r;
 	int ret = 0;
 
-	ret = tplg_load_pga(ctx, &volume);
+	ret = tplg_create_pga(ctx, &volume);
 	if (ret < 0)
 		return ret;
 
@@ -223,14 +250,14 @@ int load_pga(struct tplg_context *ctx)
 }
 
 /* load scheduler dapm widget */
-int load_pipeline(struct tplg_context *ctx)
+static int fuzzer_load_pipeline(struct tplg_context *ctx)
 {
 	struct sof_ipc_pipe_new pipeline;
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_reply r;
 	int ret;
 
-	ret = tplg_load_pipeline(ctx, &pipeline);
+	ret = tplg_create_pipeline(ctx, &pipeline);
 	if (ret < 0)
 		return ret;
 
@@ -251,14 +278,14 @@ int load_pipeline(struct tplg_context *ctx)
 }
 
 /* load src dapm widget */
-int load_src(struct tplg_context *ctx)
+static int fuzzer_load_src(struct tplg_context *ctx)
 {
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_src src = {0};
 	struct sof_ipc_comp_reply r;
 	int ret = 0;
 
-	ret = tplg_load_src(ctx, &src);
+	ret = tplg_create_src(ctx, &src);
 	if (ret < 0)
 		return ret;
 
@@ -277,14 +304,14 @@ int load_src(struct tplg_context *ctx)
 }
 
 /* load asrc dapm widget */
-int load_asrc(struct tplg_context *ctx)
+static int fuzzer_load_asrc(struct tplg_context *ctx)
 {
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_asrc asrc = {0};
 	struct sof_ipc_comp_reply r;
 	int ret = 0;
 
-	ret = tplg_load_asrc(ctx, &asrc);
+	ret = tplg_create_asrc(ctx, &asrc);
 	if (ret < 0)
 		return ret;
 
@@ -303,14 +330,14 @@ int load_asrc(struct tplg_context *ctx)
 }
 
 /* load mixer dapm widget */
-int load_mixer(struct tplg_context *ctx)
+static int fuzzer_load_mixer(struct tplg_context *ctx)
 {
 	struct fuzz *fuzzer = ctx->fuzzer;
 	struct sof_ipc_comp_mixer mixer = {0};
 	struct sof_ipc_comp_reply r;
 	int ret = 0;
 
-	ret = tplg_load_mixer(ctx, &mixer);
+	ret = tplg_create_mixer(ctx, &mixer);
 	if (ret < 0)
 		return ret;
 
@@ -329,13 +356,166 @@ int load_mixer(struct tplg_context *ctx)
 }
 
 /* load effect dapm widget */
-int load_process(struct tplg_context *ctx)
+static int fuzzer_load_process(struct tplg_context *ctx)
 {
 	return -EINVAL; /* Not implemented */
 }
 
+/* load dapm widget */
+static int fuzzer_load_widget(struct tplg_context *ctx)
+{
+	struct comp_info *temp_comp_list = ctx->info;
+	int comp_index = ctx->info_index;
+	int comp_id = ctx->comp_id;
+	int ret = 0;
+	int dev_type = ctx->dev_type;
+
+	if (!temp_comp_list) {
+		fprintf(stderr, "load_widget: temp_comp_list argument NULL\n");
+		return -EINVAL;
+	}
+
+	/* allocate memory for widget */
+	ctx->widget_size = sizeof(struct snd_soc_tplg_dapm_widget);
+	ctx->widget = malloc(ctx->widget_size);
+	if (!ctx->widget) {
+		fprintf(stderr, "error: mem alloc\n");
+		return -errno;
+	}
+
+	/* read widget data */
+	ret = fread(ctx->widget, ctx->widget_size, 1, ctx->file);
+	if (ret != 1) {
+		ret = -EINVAL;
+		goto exit;
+	}
+
+	/*
+	 * create a list with all widget info
+	 * containing mapping between component names and ids
+	 * which will be used for setting up component connections
+	 */
+	temp_comp_list[comp_index].id = comp_id;
+	temp_comp_list[comp_index].name = strdup(ctx->widget->name);
+	temp_comp_list[comp_index].type = ctx->widget->id;
+	temp_comp_list[comp_index].pipeline_id = ctx->pipeline_id;
+
+	printf("debug: loading comp_id %d: widget %s id %d\n",
+	       comp_id, ctx->widget->name, ctx->widget->id);
+
+	/* load widget based on type */
+	switch (ctx->widget->id) {
+
+	/* load pga widget */
+	case SND_SOC_TPLG_DAPM_PGA:
+		if (fuzzer_load_pga(ctx) < 0) {
+			fprintf(stderr, "error: load pga\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_AIF_IN:
+		if (load_aif_in_out(ctx, SOF_IPC_STREAM_PLAYBACK) < 0) {
+			fprintf(stderr, "error: load AIF IN failed\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_AIF_OUT:
+		if (load_aif_in_out(ctx, SOF_IPC_STREAM_CAPTURE) < 0) {
+			fprintf(stderr, "error: load AIF OUT failed\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_DAI_IN:
+		if (load_dai_in_out(ctx, SOF_IPC_STREAM_PLAYBACK) < 0) {
+			fprintf(stderr, "error: load filewrite\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_DAI_OUT:
+		if (load_dai_in_out(ctx, SOF_IPC_STREAM_CAPTURE) < 0) {
+			fprintf(stderr, "error: load filewrite\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_BUFFER:
+		if (fuzzer_load_buffer(ctx) < 0) {
+			fprintf(stderr, "error: load buffer\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_SCHEDULER:
+		/* find comp id for scheduling comp */
+		if (dev_type == FUZZER_DEV)
+			ctx->sched_id = find_widget(temp_comp_list, comp_id, ctx->widget->sname);
+
+		if (fuzzer_load_pipeline(ctx) < 0) {
+			fprintf(stderr, "error: load pipeline\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_SRC:
+		if (fuzzer_load_src(ctx) < 0) {
+			fprintf(stderr, "error: load src\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_ASRC:
+		if (fuzzer_load_asrc(ctx) < 0) {
+			fprintf(stderr, "error: load src\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_MIXER:
+		if (fuzzer_load_mixer(ctx) < 0) {
+			fprintf(stderr, "error: load mixer\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_EFFECT:
+		if (fuzzer_load_process(ctx) < 0) {
+			fprintf(stderr, "error: load effect\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	/* unsupported widgets */
+	default:
+		if (fseek(ctx->file, ctx->widget->priv.size, SEEK_CUR)) {
+			fprintf(stderr, "error: fseek unsupported widget\n");
+			ret = -errno;
+			goto exit;
+		}
+
+		printf("info: Widget type not supported %d\n", ctx->widget->id);
+		ret = tplg_create_controls(ctx->widget->num_kcontrols, ctx->file);
+		if (ret < 0) {
+			fprintf(stderr, "error: loading controls\n");
+			goto exit;
+		}
+		ret = 0;
+		break;
+	}
+
+	ret = 1;
+
+exit:
+	/* free allocated widget data */
+	free(ctx->widget);
+	return ret;
+}
+
 /* parse topology file and set up pipeline */
-int parse_topology(struct tplg_context *ctx)
+int fuzzer_parse_topology(struct tplg_context *ctx)
 {
 	struct snd_soc_tplg_hdr *hdr;
 	struct fuzz *fuzzer = ctx->fuzzer;
@@ -413,7 +593,7 @@ int parse_topology(struct tplg_context *ctx)
 			for (ctx->info_index = (ctx->info_elems - hdr->count);
 			     ctx->info_index < ctx->info_elems;
 			     ctx->info_index++) {
-				ret = load_widget(ctx);
+				ret = fuzzer_load_widget(ctx);
 				if (ret < 0) {
 					printf("error: loading widget\n");
 					goto finish;
@@ -424,7 +604,7 @@ int parse_topology(struct tplg_context *ctx)
 
 		/* set up component connections from pipeline graph */
 		case SND_SOC_TPLG_TYPE_DAPM_GRAPH:
-			if (load_graph(fuzzer, comp_list_realloc, hdr->count,
+			if (fuzzer_load_graph(fuzzer, comp_list_realloc, hdr->count,
 				       num_comps, hdr->index) < 0) {
 				fprintf(stderr, "error: pipeline graph\n");
 				return -EINVAL;
@@ -443,7 +623,7 @@ finish:
 	/* pipeline complete after pipeline connections are established */
 	for (i = 0; i < num_comps; i++)
 		if (comp_list_realloc[i].type == SND_SOC_TPLG_DAPM_SCHEDULER)
-			complete_pipeline(fuzzer, comp_list_realloc[i].id);
+			fuzzer_complete_pipeline(fuzzer, comp_list_realloc[i].id);
 
 	fprintf(stdout, "debug: %s", "topology parsing end\n");
