+++ tools/testbench/topology.c
@@ -95,65 +95,6 @@ int find_widget(struct comp_info *temp_comp_list, int count, char *name)
 	return 0;
 }
 
-/* load pipeline graph DAPM widget*/
-static int load_graph(void *dev, struct comp_info *temp_comp_list,
-		      struct testbench_prm *tp,
-		      int count, int num_comps, int pipeline_id)
-{
-	struct sof_ipc_pipe_comp_connect connection;
-	struct sof *sof = (struct sof *)dev;
-	int ret = 0;
-	int i;
-
-	for (i = 0; i < count; i++) {
-		ret = tplg_load_graph(num_comps, pipeline_id, temp_comp_list,
-				      tp->pipeline_string, &connection, tp->file, i,
-				      count);
-		if (ret < 0)
-			return ret;
-
-		/* connect source and sink */
-		if (ipc_comp_connect(sof->ipc, ipc_to_pipe_connect(&connection)) < 0) {
-			fprintf(stderr, "error: comp connect\n");
-			return -EINVAL;
-		}
-	}
-
-	/* pipeline complete after pipeline connections are established */
-	for (i = 0; i < num_comps; i++) {
-		if (temp_comp_list[i].pipeline_id == pipeline_id &&
-		    temp_comp_list[i].type == SND_SOC_TPLG_DAPM_SCHEDULER)
-			ipc_pipeline_complete(sof->ipc, temp_comp_list[i].id);
-	}
-
-	return ret;
-}
-
-/* load buffer DAPM widget */
-int load_buffer(struct tplg_context *ctx)
-{
-	struct sof *sof = ctx->sof;
-	struct sof_ipc_buffer buffer = {0};
-	int ret;
-
-	ret = tplg_load_buffer(ctx, &buffer);
-	if (ret < 0)
-		return ret;
-
-	if (tplg_load_controls(ctx->widget->num_kcontrols, ctx->file) < 0) {
-		fprintf(stderr, "error: loading controls\n");
-		return -EINVAL;
-	}
-
-	/* create buffer component */
-	if (ipc_buffer_new(sof->ipc, &buffer) < 0) {
-		fprintf(stderr, "error: buffer new\n");
-		return -EINVAL;
-	}
-
-	return 0;
-}
-
 /* load fileread component */
 static int tplg_load_fileread(struct tplg_context *ctx,
 			      struct sof_ipc_comp_file *fileread)
@@ -298,7 +239,7 @@ static int load_fileread(struct tplg_context *ctx, int dir)
 	if (ret < 0)
 		return ret;
 
-	if (tplg_load_controls(ctx->widget->num_kcontrols, file) < 0) {
+	if (tplg_create_controls(ctx->widget->num_kcontrols, file) < 0) {
 		fprintf(stderr, "error: loading controls\n");
 		return -EINVAL;
 	}
@@ -343,7 +284,7 @@ static int load_filewrite(struct tplg_context *ctx, int dir)
 	if (ret < 0)
 		return ret;
 
-	if (tplg_load_controls(ctx->widget->num_kcontrols, file) < 0) {
+	if (tplg_create_controls(ctx->widget->num_kcontrols, file) < 0) {
 		fprintf(stderr, "error: loading controls\n");
 		return -EINVAL;
 	}
@@ -393,316 +334,6 @@ int load_dai_in_out(struct tplg_context *ctx, int dir)
 		return load_filewrite(ctx, dir);
 	else
 		return load_fileread(ctx, dir);
-}
-
-/* load pda dapm widget */
-int load_pga(struct tplg_context *ctx)
-{
-	struct sof *sof = ctx->sof;
-	struct sof_ipc_comp_volume volume = {};
-	struct snd_soc_tplg_ctl_hdr *ctl = NULL;
-	struct snd_soc_tplg_mixer_control *mixer_ctl;
-	char *priv_data = NULL;
-	int32_t vol_min = 0;
-	int32_t vol_step = 0;
-	int32_t vol_maxs = 0;
-	float vol_min_db;
-	float vol_max_db;
-	int channels = 0;
-	int ret = 0;
-
-	ret = tplg_load_pga(ctx, &volume);
-	if (ret < 0)
-		return ret;
-
-	/* Only one control is supported*/
-	if (ctx->widget->num_kcontrols > 1) {
-		fprintf(stderr, "error: more than one kcontrol defined\n");
-		return -EINVAL;
-	}
-
-	/* Get control into ctl and priv_data */
-	if (ctx->widget->num_kcontrols) {
-		ret = tplg_load_one_control(&ctl, &priv_data, ctx->file);
-		if (ret < 0) {
-			fprintf(stderr, "error: failed control load\n");
-			return ret;
-		}
-
-		/* Get volume scale */
-		mixer_ctl = (struct snd_soc_tplg_mixer_control *)ctl;
-		vol_min = (int32_t)mixer_ctl->hdr.tlv.scale.min;
-		vol_step = mixer_ctl->hdr.tlv.scale.step;
-		vol_maxs = mixer_ctl->max;
-		channels = mixer_ctl->num_channels;
-	}
-
-	vol_min_db = 0.01 * vol_min;
-	vol_max_db = 0.01 * (vol_maxs * vol_step) + vol_min_db;
-	volume.min_value = round(pow(10, vol_min_db / 20.0) * 65535);
-	volume.max_value = round(pow(10, vol_max_db / 20.0) * 65536);
-	volume.channels = channels;
-
-	free(ctl);
-	free(priv_data);
-
-	/* load volume component */
-	register_comp(volume.comp.type, NULL);
-	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(&volume)) < 0) {
-		fprintf(stderr, "error: new pga comp\n");
-		return -EINVAL;
-	}
-
-	return 0;
-}
-
-/* load scheduler dapm widget */
-int load_pipeline(struct tplg_context *ctx)
-{
-	struct sof *sof = ctx->sof;
-	struct sof_ipc_pipe_new pipeline = {0};
-	FILE *file = ctx->file;
-	int ret;
-
-	ret = tplg_load_pipeline(ctx, &pipeline);
-	if (ret < 0)
-		return ret;
-
-	if (tplg_load_controls(ctx->widget->num_kcontrols, file) < 0) {
-		fprintf(stderr, "error: loading controls\n");
-		return -EINVAL;
-	}
-
-	pipeline.sched_id = ctx->sched_id;
-
-	/* Create pipeline */
-	if (ipc_pipeline_new(sof->ipc, (ipc_pipe_new *)&pipeline) < 0) {
-		fprintf(stderr, "error: pipeline new\n");
-		return -EINVAL;
-	}
-
-	return 0;
-}
-
-/* load src dapm widget */
-int load_src(struct tplg_context *ctx)
-{
-	struct sof *sof = ctx->sof;
-	struct sof_ipc_comp_src src = {0};
-	FILE *file = ctx->file;
-	int ret = 0;
-
-	ret = tplg_load_src(ctx, &src);
-	if (ret < 0)
-		return ret;
-
-	if (tplg_load_controls(ctx->widget->num_kcontrols, file) < 0) {
-		fprintf(stderr, "error: loading controls\n");
-		return -EINVAL;
-	}
-
-	/* set testbench input and output sample rate from topology */
-	if (!ctx->fs_out) {
-		ctx->fs_out = src.sink_rate;
-
-		if (!ctx->fs_in)
-			ctx->fs_in = src.source_rate;
-		else
-			src.source_rate = ctx->fs_in;
-	} else {
-		src.sink_rate = ctx->fs_out;
-	}
-
-	/* load src component */
-	register_comp(src.comp.type, NULL);
-	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(&src)) < 0) {
-		fprintf(stderr, "error: new src comp\n");
-		return -EINVAL;
-	}
-
-	return ret;
-}
-
-/* load asrc dapm widget */
-int load_asrc(struct tplg_context *ctx)
-{
-	struct snd_soc_tplg_dapm_widget *widget = ctx->widget;
-	struct sof *sof = ctx->sof;
-	struct sof_ipc_comp_asrc asrc = {0};
-	int ret = 0;
-
-	ret = tplg_load_asrc(ctx, &asrc);
-	if (ret < 0)
-		return ret;
-
-	if (tplg_load_controls(widget->num_kcontrols, ctx->file) < 0) {
-		fprintf(stderr, "error: loading controls\n");
-		return -EINVAL;
-	}
-
-	/* set testbench input and output sample rate from topology */
-	if (!ctx->fs_out) {
-		ctx->fs_out = asrc.sink_rate;
-
-		if (!ctx->fs_in)
-			ctx->fs_in = asrc.source_rate;
-		else
-			asrc.source_rate = ctx->fs_in;
-	} else {
-		asrc.sink_rate = ctx->fs_out;
-	}
-
-	/* load asrc component */
-	register_comp(asrc.comp.type, NULL);
-	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(&asrc)) < 0) {
-		fprintf(stderr, "error: new asrc comp\n");
-		return -EINVAL;
-	}
-
-	return ret;
-}
-
-static int process_init_data(struct sof_ipc_comp_process **process_ipc,
-			     struct sof_ipc_comp_process *process)
-{
-	*process_ipc = malloc(sizeof(struct sof_ipc_comp_process));
-	if (!(*process_ipc)) {
-		fprintf(stderr, "error: Failed to allocate IPC\n");
-		return -errno;
-	}
-
-	/* Copy header */
-	memcpy(*process_ipc, process, sizeof(struct sof_ipc_comp_process));
-	(*process_ipc)->size = 0;
-	return 0;
-}
-
-static int process_append_data(struct sof_ipc_comp_process **process_ipc,
-			       struct sof_ipc_comp_process *process,
-			       struct snd_soc_tplg_ctl_hdr *ctl,
-			       char *priv_data)
-{
-	int ipc_size;
-	int size = 0;
-	struct snd_soc_tplg_bytes_control *bytes_ctl;
-
-	if (*process_ipc) {
-		fprintf(stderr, "error: Only single private data append to IPC is supported\n");
-		return -EINVAL;
-	}
-
-	/* Size is process IPC plus private data minus ABI header */
-	ipc_size = sizeof(struct sof_ipc_comp_process);
-	if (ctl->ops.info == SND_SOC_TPLG_CTL_BYTES) {
-		bytes_ctl = (struct snd_soc_tplg_bytes_control *)ctl;
-		size = bytes_ctl->priv.size - sizeof(struct sof_abi_hdr);
-		ipc_size += size;
-	}
-	*process_ipc = malloc(ipc_size);
-	if (!(*process_ipc)) {
-		fprintf(stderr, "error: Failed to allocate IPC\n");
-		return -errno;
-	}
-
-	/* Copy header */
-	memcpy(*process_ipc, process, sizeof(struct sof_ipc_comp_process));
-
-	/* Copy configuration data, need to strip ABI header*/
-	memcpy((char *)*process_ipc + sizeof(struct sof_ipc_comp_process),
-	       priv_data + sizeof(struct sof_abi_hdr), size);
-	(*process_ipc)->size = size;
-	return 0;
-}
-
-/* load process dapm widget */
-int load_process(struct tplg_context *ctx)
-{
-	struct sof *sof = ctx->sof;
-	struct snd_soc_tplg_dapm_widget *widget = ctx->widget;
-	struct sof_ipc_comp_process process = {0};
-	struct sof_ipc_comp_process *process_ipc = NULL;
-	struct snd_soc_tplg_ctl_hdr *ctl = NULL;
-	struct sof_ipc_comp_ext comp_ext;
-	char *priv_data = NULL;
-	int ret = 0;
-	int i;
-
-	ret = tplg_load_process(ctx, &process, &comp_ext);
-	if (ret < 0)
-		return ret;
-
-	printf("debug uuid:\n");
-	for (i = 0; i < SOF_UUID_SIZE; i++)
-		printf("%u ", comp_ext.uuid[i]);
-	printf("\n");
-
-	/* Get control into ctl and priv_data */
-	for (i = 0; i < widget->num_kcontrols; i++) {
-		ret = tplg_load_one_control(&ctl, &priv_data, ctx->file);
-
-		if (ret < 0) {
-			fprintf(stderr, "error: failed control load\n");
-			return ret;
-		}
-
-		/* Merge process and priv_data into process_ipc */
-		if (priv_data)
-			ret = process_append_data(&process_ipc, &process, ctl, priv_data);
-
-		free(ctl);
-		free(priv_data);
-		if (ret) {
-			fprintf(stderr, "error: private data append failed\n");
-			free(process_ipc);
-			return ret;
-		}
-	}
-
-	/* Default IPC without appended data */
-	if (!process_ipc) {
-		ret = process_init_data(&process_ipc, &process);
-		if (ret)
-			return ret;
-	}
-
-	/* load process component */
-	register_comp(process_ipc->comp.type, &comp_ext);
-
-	/* Instantiate */
-	ret = ipc_comp_new(sof->ipc, ipc_to_comp_new(process_ipc));
-	free(process_ipc);
-
-	if (ret < 0)
-		fprintf(stderr, "error: new process comp\n");
-
-	return ret;
-}
-
-/* load mixer dapm widget */
-int load_mixer(struct tplg_context *ctx)
-{
-	struct sof_ipc_comp_mixer mixer = {0};
-	struct sof *sof = ctx->sof;
-	FILE *file = ctx->file;
-	int ret = 0;
-
-	ret = tplg_load_mixer(ctx, &mixer);
-	if (ret < 0)
-		return ret;
-
-	if (tplg_load_controls(ctx->widget->num_kcontrols, file) < 0) {
-		fprintf(stderr, "error: loading controls\n");
-		return -EINVAL;
-	}
-
-	/* load mixer component */
-	register_comp(mixer.comp.type, NULL);
-	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(&mixer)) < 0) {
-		fprintf(stderr, "error: new mixer comp\n");
-		return -EINVAL;
-	}
-
-	return ret;
 }
 
 static int get_next_hdr(struct tplg_context *ctx, struct snd_soc_tplg_hdr *hdr, size_t file_size)
@@ -832,8 +463,11 @@ int parse_topology(struct tplg_context *ctx)
 
 		/* set up component connections from pipeline graph */
 		case SND_SOC_TPLG_TYPE_DAPM_GRAPH:
-			if (load_graph(ctx->sof, ctx->info, tp, hdr->count,
-				       ctx->comp_id, hdr->index) < 0) {
+			if (tplg_register_graph(ctx->sof, ctx->info,
+						tp->pipeline_string,
+						tp->file, hdr->count,
+						ctx->comp_id,
+						hdr->index) < 0) {
 				fprintf(stderr, "error: pipeline graph\n");
 				ret = -EINVAL;
 				goto out;
